[CLEAN] ir_qweb: simplified and cleaned QWebcontext implementation - now using openerp.tools.safe_eval, instead of a custom eval with custom builtins - removed undefined_handler, hardcoded to a lambda function that returns None for a missing attribute
tools.safe_eval: added a parameter locals_builtins. This allows to copy the builtins
in the locals. This hack is due to the fact that the locals always returns None, allowing
to simplify templates. Otherwise we would have to test the existence of each variable
before actually using it.
However as the locals always return None for every key, the globals are never checked.
Copying the builtins inside the local allows to have a complete locals, but sligtly
break the globals/locals separation.

To be reviewed and approved.