[FIX] work around distribute_branding
distribute_branding is applied after view inheritance is applied, and does not
take it in account. As a result, altering an existing view "from the top"
(e.g. adding a div element preceding an existing div) will alter the xpath
leading to whatever element we need to deal with.

Move fakebar injection to the bottom of the body, where it can't have any
impact on preceding node paths.